requirements.txt
python-dotenv
django-allauth
psycopg2-binary
django-allauth-ui
django-widget-tweaks
Pillow